backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-allauth>=0.61.1
dj-rest-auth>=6.0.0
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=5.0.0
google-genai>=0.5.0
python-dotenv>=1.0.1
django-cors-headers>=4.3.1
dj-database-url>=2.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
